#32 refactor:account페이지의 경우 ccp 제외
